requirements.txt
cffi==2.0.0
coloredlogs==15.0.1
flatbuffers==25.9.23
humanfriendly==10.0
mpmath==1.3.0
numpy==2.0.2
onnxruntime==1.19.2
packaging==25.0
piper-tts==1.3.0
protobuf==6.33.0
pycparser==2.23
sounddevice==0.5.3
sympy==1.14.0
